# Conflicts:
#	com.vrcfury.vrcfury/Editor/VF/Utils/BlendtreeMath.cs

+++ com.vrcfury.vrcfury/Editor/VF/TmpFilePackage.cs
@@ -1,25 +1,43 @@
 using System;
+using System.Collections.Immutable;
 using System.IO;
+using System.Linq;
 using System.Reflection;
 using System.Text;
+using JetBrains.Annotations;
 using UnityEditor;
 using UnityEditor.PackageManager;
+using UnityEditor.VersionControl;
 using UnityEngine;
 using VF.Builder;
+using VF.Utils;
+using VRC.SDK3.Avatars.Components;
+using Object = System.Object;
 
 namespace VF {
     internal static class TmpFilePackage {
         private const string TmpDirPath = "Packages/com.vrcfury.temp";
         private const string TmpPackagePath = TmpDirPath + "/" + "package.json";
-        private const string LegacyTmpDirPath = "Assets/_VRCFury";
-        private const string LegacyPrefabsImportedMarker = TmpDirPath + "/LegacyPrefabsImported";
 
         public static void Cleanup() {
-            var tmpDir = GetPath();
+            var tmpDir = GetPathNullable();
+            if (tmpDir == null) return;
+            
+            var usedFolders = Resources.FindObjectsOfTypeAll<VRCAvatarDescriptor>()
+                .SelectMany(VRCAvatarUtils.GetAllControllers)
+                .Where(c => !c.isDefault && c.controller != null)
+                .Select(c => AssetDatabase.GetAssetPath(c.controller))
+                .Where(path => !string.IsNullOrEmpty(path))
+                .Select(VRCFuryAssetDatabase.GetDirectoryName)
+                .ToImmutableHashSet();
+
             VRCFuryAssetDatabase.WithAssetEditing(() => {
                 VRCFuryAssetDatabase.DeleteFiltered(tmpDir, path => {
+                    if (usedFolders.Any(used => path.StartsWith($"{used}/") || path == used || used.StartsWith($"{path}/"))) return false;
                     if (path.StartsWith(tmpDir + "/SPS")) return false;
+                    if (path.StartsWith(tmpDir + "/XR")) return false;
                     if (path.StartsWith(tmpDir + "/package.json")) return false;
+                    if (path.StartsWith(tmpDir + "/PlayModeSettings")) return false;
                     if (path.StartsWith(tmpDir + "/LegacyPrefabsImported")) return false;
                     return true;
                 });
@@ -31,16 +49,23 @@ namespace VF {
             VRCFuryAssetDatabase.WithoutAssetEditing(() => {});
         }
 
+        [CanBeNull]
+        public static string GetPathNullable() {
+            if (!AssetDatabase.IsValidFolder(TmpDirPath)) return null;
+            return TmpDirPath;
+        }
+
         public static string GetPath() {
-            var importLegacyPrefabs = false;
-            if ((Directory.Exists(LegacyTmpDirPath) || Directory.Exists(TmpDirPath)) &&
-                !File.Exists(LegacyPrefabsImportedMarker)) {
-                importLegacyPrefabs = true;
-            }
+            var tmpDir = GetPathNullable();
+            if (tmpDir == null) throw new Exception("VRCFury Temp Files package has not been created yet. Try again?");
+            return tmpDir;
+        }
+
+        private static void InitIfMissing() {
+            if (GetPathNullable() != null) return;
 
             if (!Directory.Exists(TmpDirPath)) {
                 Directory.CreateDirectory(TmpDirPath); 
-                File.Create(LegacyPrefabsImportedMarker).Close();
             }
 
             if (!File.Exists(TmpPackagePath) ||
@@ -49,13 +74,6 @@ namespace VF {
 
                 EditorApplication.delayCall += ReresolvePackages;
             }
-
-            if (importLegacyPrefabs) {
-                LegacyPrefabUnpacker.ScanOnce();
-                File.Create(LegacyPrefabsImportedMarker).Close();
-            }
-
-            return TmpDirPath;
         }
 
         private static void ReresolvePackages() {
@@ -70,7 +88,7 @@ namespace VF {
 
         [InitializeOnLoadMethod]
         private static void Init() {
-            GetPath();
+            Scheduler.Schedule(InitIfMissing, 5000);
         }
 
         private static readonly string PackageJson =
